feat(RTC): 添加RTC模块支持并优化电机控制逻辑
启用RTC模块并更新相关配置，修改电机控制逻辑以解决符号转换警告。调整M3508电机数组顺序，优化CAN通信接口，移除冗余结构体定义。更新构建配置以包含RTC相关文件。

修复DM_Motor.c中的符号转换警告，调整电机控制逻辑中的角度计算方式。重构BSP_Can.c中的CAN发送接口，移除不必要的函数指针结构体。更新项目配置以启用RTC模块，并调整系统时钟初始化。

更新构建日志和依赖文件，确保所有修改正确编译。调整FreeRTOS任务优先级，优化系统初始化流程。修改GPIO控制逻辑，将其移至初始化任务中执行。

+++ Hero_board2/DEV/Inc/DM_Motor.h
@@ -31,10 +31,10 @@ typedef struct
   uint8_t temperatureRotor;
   float torqueInit;
   float torque;
-  float angleInit;
+  uint32_t angleInit;
   float speedInit;
 
-  uint16_t lastAngle;
+  uint32_t lastAngle;
 
   int32_t targetSpeed;
   int32_t targetAngle;
